feat(claude-tools): replace gh add-sub-issue with add-sub-issues for multiple sub-issue support
via [Happy](https://happy.engineering)
