+++ packages/claude-tools/src/commands/gh/index.ts
@@ -1,4 +1,4 @@
 export const commands: Record<string, () => Promise<void>> = {
-  "add-sub-issue": () => import("./add-sub-issue").then((m) => m.main()),
+  "add-sub-issues": () => import("./add-sub-issues").then((m) => m.main()),
   "list-sub-issues": () => import("./list-sub-issues").then((m) => m.main()),
 };
